Stream library fixes (#75)
* Fixed bug in StreamArbiter

With the previous version, it was possible for the arbiter to switch source stream before observing the last transfer.

* Add support for null transfers in StreamSerializer

This applies to the case where the input count is wide enough to represent zero, as opposed to 0 representing the maximum count as a special case. The serializer used to always handle 0 as 2**count_width.

* Added proper support for MAX_COUNT in StreamElementCounter

It used to ignore this generic. Also, rollover behavior was inconsistently implemented, leading to possible cases where the total count does not match the element count. StreamElementCounter now behaves as specified in the interface comments, also supporting non-power-of-two MAX_COUNTs.