Fletchgen redesign (#98)
Major changes:

* A new library called Cerata is introduced.

Cerata is a library that helps to programmatically set up hardware structures with complex types. 
It allows you to set up structures as graphs, and then select some back-end to output the graph as an HDL design. It supports complex, nested types (e.g. records of streams) and transformations of graphs (e.g. inserting profiling components etc.) and several back-ends (at this moment VHDL and DOT), and a sort of "reflection" where component models could implement functions that operate on already generated designs.
Cerata is heavily used in the new Fletchgen (see next). for now we will consider this library as a component of Fletcher, but it is not and should not depend on any of the components of Fletcher, so that it may become a project on its own later on.

* A new version of Fletchgen based on Cerata.

The old Fletchgen was too messy and monolithic. Basing it on something more generic like Cerata helps to modularize the design. This allows for easier debugging, integration and extension. It now generates more levels of hierarchy in the design to be logically consistent with the view of the data structure that Arrow Schemas and Recordbatches can convey.

Some leftover work that is spawned through this pull request, but should be done for Release 0.1 anyway:

* Some changes to how a kernel should be implemented are reflected in this update as well. This invalidates all existing examples. I would like to solve them as separate pull requests. It might even be worthwhile to consider separating the examples from the project as well, as they are sort of hard to continuously maintain and test as they need to be synthesized, placed, routed, etc... Even if it would be automated through CI, every example increases CI run-time with at least several hours. It's probably better to set them up in (a) seperate reposit(y/ies) and connect them to some tag (or future release) of the project.

* Revisit all existing documentation