Restructure cerata to provision auto clock domain crossing in the future
This restructures the code quite a bit, since the previous CDC constructs/stubs were attached to types, which doesn't make a lot of sense. I've introduced a Synchronous class that holds information about nodes that operate in a synchronous environment. For now this just holds a clock domain. The clock domain in turn doesn't hold much either yet, just a name, but I think the constructs are placed more conveniently for the future w.r.t all Cerata abstractions. For now if you try to attach two signal or port nodes with different clock domains, only a warning is generated.